Moved viewmodels to their own folder
