+++ YousifsSolution/YousifsProject/Views/Houses/ViewModels/EditHouseVM.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc.Rendering;
 using System.ComponentModel.DataAnnotations;
 
-namespace YousifsProject.Views.Houses
+namespace YousifsProject.Views.Houses.ViewModels
 {
-    public class BuildHouseVM
+    public class EditHouseVM
     {
         [StringLength(maximumLength: 50, MinimumLength = 4, ErrorMessage = "The address must be at least 4 letters")]
         [Required(ErrorMessage = "Write an address")]
@@ -33,6 +33,9 @@ namespace YousifsProject.Views.Houses
         [Required]
         public string TypeOfRoof { get; set; }
 
+        public int Id { get; set; }
+
+        public string[]? TypeOfRoofsArray { get; set; }
         public SelectListItem[]? FloorArray { get; set; }
     }
 }
