+++ YousifsSolution/YousifsProject/Views/_ViewImports.cshtml
@@ -1,4 +1,7 @@
 ﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @using YousifsProject.Views.Houses;
 @using YousifsProject.Views.Identity;
+@using YousifsProject.Views.Houses.ViewModels
+@using YousifsProject.Views.Identity.ViewModels;
+@using YousifsProject.Controllers;
